requirements.txt
APScheduler==3.9.1
discord.py==2.2.3
emoji==2.0.0
python-dotenv==1.0.0
aiosqlite==0.18.0
